DietItem refactoring
